Initial stage of CProver update to 5.8

+++ trunk/cprover/src/util/file_util.cpp
@@ -8,7 +8,15 @@ Date: January 2012
 
 \*******************************************************************/
 
+/// \file
+/// File Utilities
+
+#include "file_util.h"
+
+#include "invariant.h"
+
 #include <cerrno>
+#include <cstring>
 
 #if defined(__linux__) || \
     defined(__FreeBSD_kernel__) || \
@@ -31,58 +39,31 @@ Date: January 2012
 #define chdir _chdir
 #define popen _popen
 #define pclose _pclose
-#else
-#include <cstring>
 #endif
 
-#include "file_util.h"
-
-/*******************************************************************\
-
-Function: get_current_working_directory
-
-  Inputs: none
-
- Outputs: current working directory
-
- Purpose:
-
-\*******************************************************************/
-
+/// \return current working directory
 std::string get_current_working_directory()
 {
-  unsigned bsize=50;
-
-  char *buf=reinterpret_cast<char*>(malloc(sizeof(char)*bsize));
-  if(!buf)
-    abort();
-
+  #ifndef _WIN32
   errno=0;
+  char *wd=realpath(".", nullptr);
+  INVARIANT(
+    wd!=nullptr && errno==0,
+    std::string("realpath failed: ")+strerror(errno));
 
-  while(buf && getcwd(buf, bsize-1)==NULL && errno==ERANGE)
-  {
-    bsize*=2;
-    buf=reinterpret_cast<char*>(realloc(buf, sizeof(char)*bsize));
-  }
-
-  std::string working_directory=buf;
-  free(buf);
+  std::string working_directory=wd;
+  free(wd);
+  #else
+  char buffer[4096];
+  DWORD retval=GetCurrentDirectory(4096, buffer);
+  CHECK_RETURN(retval>0);
+  std::string working_directory(buffer);
+  #endif
 
   return working_directory;
 }
 
-/*******************************************************************\
-
-Function: delete_directory
-
-  Inputs: path
-
- Outputs:
-
- Purpose: deletes all files in 'path' and then the directory itself
-
-\*******************************************************************/
-
+/// deletes all files in 'path' and then the directory itself
 #ifdef _WIN32
 
 void delete_directory_utf16(const std::wstring &path)
@@ -117,10 +98,10 @@ void delete_directory(const std::string &path)
   delete_directory_utf16(utf8_to_utf16_little_endian(path));
 #else
   DIR *dir=opendir(path.c_str());
-  if(dir!=NULL)
+  if(dir!=nullptr)
   {
     struct dirent *ent;
-    while((ent=readdir(dir))!=NULL)
+    while((ent=readdir(dir))!=nullptr)
     {
       // Needed for Alpine Linux
       if(strcmp(ent->d_name, ".")==0 || strcmp(ent->d_name, "..")==0)
@@ -129,12 +110,18 @@ void delete_directory(const std::string &path)
       std::string sub_path=path+"/"+ent->d_name;
 
       struct stat stbuf;
-      stat(sub_path.c_str(), &stbuf);
+      int result=stat(sub_path.c_str(), &stbuf);
+      if(result!=0)
+        throw std::string("Stat failed: ")+std::strerror(errno);
 
       if(S_ISDIR(stbuf.st_mode))
         delete_directory(sub_path);
       else
-        remove(sub_path.c_str());
+      {
+        result=remove(sub_path.c_str());
+        if(result!=0)
+          throw std::string("Remove failed: ")+std::strerror(errno);
+      }
     }
     closedir(dir);
   }
@@ -142,19 +129,8 @@ void delete_directory(const std::string &path)
 #endif
 }
 
-/*******************************************************************\
-
-Function: concat_dir_file
-
-  Inputs: directory name and file name
-
- Outputs: concatenation of directory and file, if the file path is
-          relative
-
- Purpose:
-
-\*******************************************************************/
-
+/// \par parameters: directory name and file name
+/// \return concatenation of directory and file, if the file path is relative
 std::string concat_dir_file(
   const std::string &directory,
   const std::string &file_name)
